+++ onnxruntime/core/dlpack/dlpack_converter.h
@@ -5,7 +5,7 @@
 
 #pragma once
 
-#include "core/framework/ml_value.h"
+#include "core/framework/ort_value.h"
 #include <dlpack/dlpack.h>
 
 namespace onnxruntime {
