+++ onnxruntime/core/framework/memory_info.cc
@@ -1,7 +1,7 @@
 #if !defined(ORT_MINIMAL_BUILD) && defined(ORT_MEMORY_PROFILE)
 #include "core/framework/memory_info.h"
 #include "core/framework/mem_pattern.h"
-#include "core/framework/ml_value.h"
+#include "core/framework/ort_value.h"
 
 #include <fstream>
 #include <numeric>
