+++ onnxruntime/core/framework/node_index_info.h
@@ -6,7 +6,7 @@
 #include <vector>
 
 #include "core/common/common.h"
-#include "core/framework/ml_value.h"
+#include "core/framework/ort_value.h"
 #include "core/graph/basic_types.h"
 
 namespace onnxruntime {
