+++ onnxruntime/core/framework/sequential_executor.h
@@ -10,7 +10,7 @@
 #include "core/common/logging/logging.h"
 #include "core/framework/iexecutor.h"
 #include "core/framework/framework_common.h"
-#include "core/framework/ml_value.h"
+#include "core/framework/ort_value.h"
 #include "core/framework/session_state.h"
 #include "core/graph/graph_viewer.h"
 #include "core/framework/op_kernel_context_internal.h"
