+++ onnxruntime/core/framework/tensor.cc
@@ -7,7 +7,7 @@
 #include "core/common/safeint.h"
 #include "core/framework/allocatormgr.h"
 #include "core/framework/data_types.h"
-#include "core/framework/ml_value.h"
+#include "core/framework/ort_value.h"
 #include "core/framework/utils.h"
 
 namespace onnxruntime {
