+++ onnxruntime/core/optimizer/optimizer_execution_frame.h
@@ -10,7 +10,7 @@
 #include "core/framework/data_transfer_manager.h"
 #include "core/framework/execution_frame.h"
 #include "core/framework/ort_value_name_idx_map.h"
-#include "core/framework/ml_value.h"
+#include "core/framework/ort_value.h"
 #include "core/framework/callback.h"
 
 namespace onnxruntime {
