+++ onnxruntime/core/providers/cuda/controlflow/scan.cc
@@ -6,7 +6,7 @@
 
 #include "core/providers/cuda/cuda_common.h"
 #include "core/providers/cuda/tensor/transpose.h"
-#include "core/framework/ml_value.h"
+#include "core/framework/ort_value.h"
 
 // TODO: It's ugly to include a .cc file but this .cc file defines the implementation of some templates which we need.
 #include "core/framework/ort_value_tensor_slicer.cc"
