+++ onnxruntime/core/session/IOBinding.h
@@ -9,7 +9,7 @@
 #include "core/framework/execution_provider.h"
 #include "core/common/status.h"
 #include "core/graph/basic_types.h"
-#include "core/framework/ml_value.h"
+#include "core/framework/ort_value.h"
 #include "core/session/inference_session.h"
 #include "core/common/logging/logging.h"
