+++ onnxruntime/test/framework/test_utils.h
@@ -8,7 +8,7 @@
 #include "core/framework/allocatormgr.h"
 #include "core/framework/execution_provider.h"
 #include "core/providers/cpu/cpu_execution_provider.h"
-#include "core/framework/ml_value.h"
+#include "core/framework/ort_value.h"
 
 #include "gsl/gsl"
