+++ orttraining/orttraining/core/framework/torch/torch_proxy.h
@@ -6,7 +6,7 @@
 #include "orttraining/core/framework/torch/python_common.h"
 
 #ifndef SHARED_PROVIDER
-#include "core/framework/ml_value.h"
+#include "core/framework/ort_value.h"
 #include "core/framework/op_kernel_context_internal.h"
 #include "core/platform/env.h"
 #endif
