+++ orttraining/orttraining/core/graph/optimizer/adam_optimizer_builder.cc
@@ -4,7 +4,7 @@
 #include "orttraining/core/graph/optimizer/adam_optimizer_builder.h"
 #include "orttraining/core/graph/graph_augmenter.h"
 #include "core/util/math.h"
-#include "core/framework/ml_value.h"
+#include "core/framework/ort_value.h"
 #include "core/framework/tensorprotoutils.h"
 
 namespace onnxruntime {
