+++ orttraining/orttraining/core/graph/optimizer/lamb_optimizer_builder.cc
@@ -5,7 +5,7 @@
 #include "orttraining/core/graph/optimizer/lamb_optimizer_builder.h"
 #include "orttraining/core/graph/optimizer_builder.h"
 #include "orttraining/core/graph/graph_augmenter.h"
-#include "core/framework/ml_value.h"
+#include "core/framework/ort_value.h"
 #include "core/framework/tensorprotoutils.h"
 #include "core/util/math.h"
 #include "onnx/defs/attr_proto_util.h"
