+++ orttraining/orttraining/core/session/tensor_helper.h
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-#include "core/framework/ml_value.h"
+#include "core/framework/ort_value.h"
 #include "core/session/inference_session.h"
 
 namespace onnxruntime {
