+++ orttraining/orttraining/eager/ort_aten.h
@@ -4,7 +4,7 @@
 #pragma once
 
 #include <torch/extension.h>
-#include <core/framework/ml_value.h>
+#include <core/framework/ort_value.h>
 
 #include "ort_util.h"
 #include "ort_ops.h"
