+++ orttraining/orttraining/eager/ort_ops.h
@@ -3,7 +3,7 @@
 
 #pragma once
 
-#include <core/framework/ml_value.h>
+#include <core/framework/ort_value.h>
 #include <core/eager/ort_kernel_invoker.h>
 
 namespace torch_ort {
