+++ orttraining/orttraining/models/runner/data_loader.h
@@ -10,7 +10,7 @@
 #include "core/common/common.h"
 #include "core/common/path_string.h"
 #include "core/framework/framework_common.h"
-#include "core/framework/ml_value.h"
+#include "core/framework/ort_value.h"
 #include "core/platform/path_lib.h"
 #include "core/platform/threadpool.h"
 #include "orttraining/models/runner/training_util.h"
